[#286] feat: 게시글 삭제 시 confirm 창 활성화

+++ frontend/src/components/articles/ArticleDetailButtons.jsx
@@ -6,21 +6,23 @@ import style from "style/articles/ArticleDetailButtons.module.css";
 
 function ArticleDetailButtons({ article, articleId }) {
 	const SERVER_URL = process.env.REACT_APP_SERVER_URL;
-	const username = useSelector((state) => state.userInfo.username);
+	const usermail = useSelector((state) => state.userInfo.email);
 	const navigate = useNavigate();
 
-	const handleDeleteButtonClick = () => {
-		axios({
-			url: `${SERVER_URL}/api/boards/articles/${articleId}`,
-			method: "delete",
-		}).then(() => {
-			navigate(-1);
-		});
+	const handleDeleteButtonClick = async () => {
+		if (window.confirm("글을 삭제 하시겠습니까?")) {
+			axios({
+				url: `${SERVER_URL}/api/boards/articles/${articleId}`,
+				method: "delete",
+			}).then(() => {
+				navigate(-1);
+			});
+		}
 	};
 
 	return (
 		<div className={style.button_box}>
-			{article.username === username ? (
+			{article.email === usermail ? (
 				<div>
 					<Link
 						to={`/articles/${articleId}/update`}
